refactor: userSearchPopUp
